requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
billiard==4.2.4
boto3==1.42.63
botocore==1.42.63
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.135.1
gunicorn==21.2.0
flake8==7.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jmespath==1.1.0
joblib==1.5.3
kombu==5.6.2
librt==0.8.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
MouseInfo==0.1.3
mypy==1.19.1
mypy_extensions==1.1.0
newsapi-python==0.2.7
nltk==3.9.3
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pathspec==1.0.4
pillow==12.1.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyMuPDF==1.27.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytweening==1.2.0
redis==7.3.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rsa==4.9.1
rubicon-objc==0.5.3
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
textblob==0.19.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
